Demo_News_RAG_Neo4j_Version_1/requirements.txt



# ######### version #########
# langchain==0.3.25
# langchain-community==0.3.25
# langchain-google-genai==2.1.5 
# langchain-neo4j==0.4.0
# langchain-experimental==0.3.4
# python-dotenv==1.1.0
# openai==1.88.0
# #streamlit==1.46.0

langchain==0.3.25 
langchain-community==0.3.25 
langchain-google-genai==2.1.5 
langchain-neo4j==0.4.0 
langchain-experimental==0.3.4 
python-dotenv==1.0.1 
openai==1.18.0
streamlit==1.32.2
